+++ src/main/java/com/gulbi/Backend/domain/contract/application/dto/ApplicationStatusProjection.java
@@ -1,0 +1,9 @@
+package com.gulbi.Backend.domain.contract.application.dto;
+
+import java.time.LocalDate;
+
+public interface ApplicationStatusProjection {
+	LocalDate getReservationDate();
+	Integer getHasReserving();
+	Integer getHasReservingOrUsing();
+}
