+++ src/main/java/com/gulbi/Backend/domain/contract/application/exception/temp.java
@@ -1,0 +1,4 @@
+package com.gulbi.Backend.domain.contract.application.exception;
+
+public class temp {
+}
