--- src/main/java/com/gulbi/Backend/domain/contract/dto/LenderApprovalCommand.java
@@ -1,19 +1,0 @@
-package com.gulbi.Backend.domain.contract.dto;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import lombok.Getter;
-
-@Getter
-public class LenderApprovalCommand {
-	private MultipartFile finalContractImage;
-	private Long contractId;
-
-	public LenderApprovalCommand() {
-	}
-
-	public LenderApprovalCommand(MultipartFile finalContractImage, Long contractId) {
-		this.finalContractImage = finalContractImage;
-		this.contractId = contractId;
-	}
-}
