--- src/main/java/com/gulbi/Backend/domain/rental/application/dto/ApplicationStatusProjection.java
@@ -1,9 +1,0 @@
-package com.gulbi.Backend.domain.rental.application.dto;
-
-import java.time.LocalDate;
-
-public interface ApplicationStatusProjection {
-	LocalDate getReservationDate();
-	Integer getHasReserving();
-	Integer getHasReservingOrUsing();
-}
