--- src/main/java/com/gulbi/Backend/domain/rental/application/exception/temp.java
@@ -1,4 +1,0 @@
-package com.gulbi.Backend.domain.rental.application.exception;
-
-public class temp {
-}
