+++ src/main/java/com/gulbi/Backend/domain/rental/product/service/product/crud/ProductCrudServiceImpl.java
@@ -208,6 +208,7 @@ public class ProductCrudServiceImpl implements ProductCrudService {
                 .className(className)
                 .responseApiCode(ProductErrorCode.PRODUCT_NOT_FOUND)
                 .build();
+
         throw new ProductException.ProductNotFoundException(exceptionMetaData);
     }
