+++ src/main/java/com/gulbi/Backend/global/error/ErrorMetaDataFactory.java
@@ -1,0 +1,4 @@
+package com.gulbi.Backend.global.error;
+
+public class ErrorMetaDataFactory {
+}
