+++ src/main/java/com/gulbi/Backend/global/error/InfrastructureException.java
@@ -1,0 +1,6 @@
+package com.gulbi.Backend.global.error;
+
+public class InfrastructureException extends RuntimeException{
+	private ExceptionMetaData exceptionMetaData;
+
+}
